+++ CHANGELOG.md
@@ -65,6 +65,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   getting and setting the currently selected file.
 - Add `noFilePicker` property (`no-file-picker` attribute) to
   `<code-sample-editor>` which disables the top file selection tab-bar.
+- Add `noLineNumbers` property (`no-line-numbers` attribute) to
+  `<code-sample-editor>` which disables the left-hand-side gutter with line
+  numbers.
 
 ### Fixed